chapter5/rag-agent/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp==1.25.0",
    "nest-asyncio==1.6.0",
    "python-dotenv>=1.2.1",
    "strands-agents==1.20.0",
    "streamlit==1.52.2",
]
